--- web_flask/app/__init__.py
@@ -1,8 +1,0 @@
-from flask import Flask
-
-def create_app():
-    app = Flask(__name__)
-
-    with app.app_context():
-        from . import routes
-        return app
