+++ web_flask/static/styles/home.css
@@ -1,0 +1,64 @@
+body {
+	font-family: Arial, sans-serif;
+	margin: 0;
+	padding: 0;
+	display: flex;
+	height: 100vh;
+	overflow: hidden;
+}
+.container {
+	display: flex;
+	width: 100%;
+	height: 100%;
+}
+.notes-section, .chat-section {
+	padding: 10px;
+}
+.notes-section {
+	width: 70%;
+	border-right: 1px solid #ccc;
+	overflow-y: auto;
+}
+.chat-section {
+	width: 30%;
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+}
+.note, .chat-message {
+	border: 1px solid #ccc;
+	margin-bottom: 10px;
+	padding: 10px;
+	border-radius: 5px;
+}
+.note-title {
+	font-weight: bold;
+	margin-bottom: 5px;
+}
+.chat-box {
+	display: flex;
+	align-items: center;
+}
+.chat-input {
+	flex: 1;
+	padding: 10px;
+	border: 1px solid #ccc;
+	border-radius: 5px;
+	margin-right: 10px;
+}
+.add-note-btn {
+	position: absolute;
+	bottom: 20px;
+	left: 20px;
+	width: 50px;
+	height: 50px;
+	background-color: #007bff;
+	color: white;
+	border: none;
+	border-radius: 50%;
+	font-size: 30px;
+	cursor: pointer;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
