+++ web_flask/templates/home.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<link rel="stylesheet" href="../static/styles/home.css">
+    <title>TakingNote</title>
+</head>
+<body>
+    <div class="container">
+        <div class="notes-section">
+            <div class="note">
+                <div class="note-title">Title 1</div>
+                <div class="note-content">Note content goes here...</div>
+            </div>
+            <div class="note">
+                <div class="note-title">Title 2</div>
+                <div class="note-content">Note content goes here...</div>
+            </div>
+            <div class="note">
+                <div class="note-title">Title 3</div>
+                <div class="note-content">Note content goes here...</div>	
+            </div>
+			<button class="add-note-btn">+</button>
+        </div>
+        <div class="chat-section">
+            <div class="chat-messages">
+                <div class="chat-message"><strong>You:</strong> Hello</div>
+                <div class="chat-message"><strong>AI:</strong> Hi there!</div>
+                <div class="chat-message"><strong>You:</strong> How are you?</div>
+            </div>
+            <div class="chat-box">
+                <input type="text" class="chat-input" placeholder="Ask AI">
+            </div>
+        </div>
+    </div>
+</body>
+</html>
+
